refactor(migrations): Extract migration analysis into package
Package structure:
- posthog/management/migration_analysis/
  - models.py: RiskLevel, OperationRisk, MigrationRisk
  - operations.py: All operation-specific analyzers
  - analyzer.py: RiskAnalyzer and combination checking
  - discovery.py: MigrationDiscovery utilities (was migration_utils)
  - test/: Unit tests for the package

Benefits:
- Clean separation of concerns
- Easier to extend with new analyzers
- Better testability
- Reusable components

All tests passing, command works unchanged.
