+++ posthog/management/migration_analysis/__init__.py
@@ -1,0 +1,18 @@
+"""
+Migration analysis package for Django migrations.
+
+Provides risk analysis and validation tools for database migrations.
+"""
+
+from posthog.management.migration_analysis.analyzer import RiskAnalyzer
+from posthog.management.migration_analysis.discovery import MigrationDiscovery, MigrationInfo
+from posthog.management.migration_analysis.models import MigrationRisk, OperationRisk, RiskLevel
+
+__all__ = [
+    "MigrationRisk",
+    "OperationRisk",
+    "RiskLevel",
+    "RiskAnalyzer",
+    "MigrationDiscovery",
+    "MigrationInfo",
+]
