+++ posthog/management/migration_analysis/test/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for migration analysis package."""
